+++ Framework/Rendering/SpriteRenderer.cs
@@ -12,7 +12,9 @@
     /// </summary>
     public class SpriteRenderer : BaseComponent, IDrawableComponent, IAssetComponent<Sprite>, IRotatable {
         private readonly ResettableLazy<BoundingArea> _boundingArea;
+        private readonly ResettableLazy<Transform> _pixelTransform;
         private readonly ResettableLazy<RotatableTransform> _rotatableTransform;
+        private bool _snapToPixels;
         private Sprite _sprite;
 
         /// <summary>
@@ -20,6 +22,7 @@
         /// </summary>
         public SpriteRenderer() {
             this._boundingArea = new ResettableLazy<BoundingArea>(this.CreateBoundingArea);
+            this._pixelTransform = new ResettableLazy<Transform>(this.CreatePixelTransform);
             this._rotatableTransform = new ResettableLazy<RotatableTransform>(this.CreateRotatableTransform);
         }
 
@@ -49,6 +52,33 @@
         public Rotation Rotation { get; private set; } = new Rotation();
 
         /// <summary>
+        /// Gets or sets a value indicating whether this sprite renderer should snap to the pixel
+        /// ratio defined in <see cref="IGameSettings"/>.
+        /// </summary>
+        /// <remarks>Snapping to pixels will disable rotations on this renderer.</remarks>
+        /// <value><c>true</c> if this should snap to pixels; otherwise, <c>false</c>.</value>
+        [DataMember]
+        public bool SnapToPixels {
+            get {
+                return this._snapToPixels;
+            }
+
+            set {
+                if (value != this._snapToPixels) {
+                    this._snapToPixels = value;
+                    if (!this._snapToPixels) {
+                        this._rotatableTransform.Reset();
+                    }
+                    else {
+                        this._pixelTransform.Reset();
+                    }
+
+                    this._boundingArea.Reset();
+                }
+            }
+        }
+
+        /// <summary>
         /// Gets or sets the sprite.
         /// </summary>
         /// <value>The sprite.</value>
@@ -72,7 +102,12 @@
 
         /// <inheritdoc/>
         public void Draw(GameTime gameTime, BoundingArea viewBoundingArea) {
-            MacabreGame.Instance.SpriteBatch.Draw(this.Sprite, this._rotatableTransform.Value, this.Color);
+            if (this._snapToPixels) {
+                MacabreGame.Instance.SpriteBatch.Draw(this.Sprite, this._pixelTransform.Value, this.Color);
+            }
+            else {
+                MacabreGame.Instance.SpriteBatch.Draw(this.Sprite, this._rotatableTransform.Value, this.Color);
+            }
         }
 
         /// <inheritdoc/>
@@ -142,18 +177,32 @@
                     this.GetWorldTransform(offset + new Vector2(0f, height), rotationAngle).Position
                 };
 
-                var minimumX = points.Min(x => x.X);
-                var minimumY = points.Min(x => x.Y);
-                var maximumX = points.Max(x => x.X);
-                var maximumY = points.Max(x => x.Y);
+                if (this.SnapToPixels) {
+                    var minimumX = points.Min(x => x.X).ToPixelSnappedValue();
+                    var minimumY = points.Min(x => x.Y).ToPixelSnappedValue();
+                    var maximumX = points.Max(x => x.X).ToPixelSnappedValue();
+                    var maximumY = points.Max(x => x.Y).ToPixelSnappedValue();
 
-                result = new BoundingArea(new Vector2(minimumX, minimumY), new Vector2(maximumX, maximumY));
+                    result = new BoundingArea(new Vector2(minimumX, minimumY), new Vector2(maximumX, maximumY));
+                }
+                else {
+                    result = new BoundingArea(new Vector2(points.Min(x => x.X), points.Min(x => x.Y)), new Vector2(points.Max(x => x.X), points.Max(x => x.Y)));
+                }
             }
             else {
                 result = new BoundingArea();
             }
 
             return result;
+        }
+
+        private Transform CreatePixelTransform() {
+            var worldTransform = this.GetWorldTransform(this.Offset.Amount * GameSettings.Instance.InversePixelsPerUnit);
+            var pixelsPerUnit = GameSettings.Instance.PixelsPerUnit;
+            var inversePixelsPerUnit = GameSettings.Instance.InversePixelsPerUnit;
+            var position = new Vector2((int)Math.Round(worldTransform.Position.X * pixelsPerUnit, 0) * inversePixelsPerUnit, (int)Math.Round(worldTransform.Position.Y * pixelsPerUnit, 0) * inversePixelsPerUnit);
+            var scale = new Vector2((int)Math.Round(worldTransform.Scale.X, 0), (int)Math.Round(worldTransform.Scale.Y));
+            return new Transform(position, scale);
         }
 
         private RotatableTransform CreateRotatableTransform() {
@@ -170,6 +219,7 @@
         }
 
         private void Offset_AmountChanged(object sender, EventArgs e) {
+            this._pixelTransform.Reset();
             this._rotatableTransform.Reset();
             this._boundingArea.Reset();
         }
@@ -182,6 +232,7 @@
 
         private void Self_TransformChanged(object sender, EventArgs e) {
             this._boundingArea.Reset();
+            this._pixelTransform.Reset();
             this._rotatableTransform.Reset();
         }
     }